requirements.txt
google-genai>=0.3.0
pillow>=9.0.0
PyPDF2>=3.0.0
requests>=2.25.0